Install Jest and add integration test for signup
I've setup Jest for use with a TypeScript Express
server as well as adding SuperTest for testing
HTTP methods. At the moment the signup route is
tested. The prismaClient has also been updated to
switch between a development database and test
database when running tests. The test file will
clear the database each time the tests are run as
well as mock sending actual verification emails.