Fix migrate and add failing GET conversation test
The previous script for updating the database was
only applying migrations to the development database
and not the test database as well. Consequently
tests were failing as a result. The new script
applies migrations to the test and development
databases. Furthermore, jest has had it's maximum
workers set to 1 to prevent race conditions when
updating the database, and a failing test has been
added for retrieving users conversations on the
/conversations route