fix: Remove friends and friendsOf fields from returned user data
Friends are a seperate resource than user data.
When a logged in user needs to see if someone is
their friend they now just compare their friendsOf
list of usernames to the list of all usernames.
It was unneccesary to return information about all
users friends and who they are friendsOf